+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
+## [3.0.0-rc.1] - 2025-10-01
 
 ### Changed
 
@@ -185,7 +185,7 @@ _Stable release based on [0.1.0-rc.1]._
 
 - New changelog file.
 
-[Unreleased]: https://https://github.com/internetguru/laravel-feedback/compare/staging...dev
+[3.0.0-rc.1]: https://github.com/internetguru/laravel-feedback/releases/tag/v2.1.3
 [2.1.3]: https://https://github.com/internetguru/laravel-feedback/compare/v2.1.2...v2.1.3
 [2.1.2]: https://https://github.com/internetguru/laravel-feedback/compare/v2.1.1...v2.1.2
 [2.1.1]: https://https://github.com/internetguru/laravel-feedback/compare/v2.1.0...v2.1.1